extracted common traits
